+++ src/test/java/com/maxmind/minfraud/response/EmailTest.java
@@ -3,6 +3,8 @@ package com.maxmind.minfraud.response;
 import com.fasterxml.jackson.jr.ob.JSON;
 import org.junit.Test;
 
+import java.time.LocalDate;
+
 import static org.junit.Assert.*;
 
 public class EmailTest extends AbstractOutputTest {
@@ -14,6 +16,9 @@ public class EmailTest extends AbstractOutputTest {
                 JSON.std
                         .composeString()
                         .startObject()
+                        .startObjectField("domain")
+                        .put("first_seen", "2014-02-03")
+                        .end()
                         .put("is_disposable", false)
                         .put("is_free", false)
                         .put("is_high_risk", true)
@@ -22,10 +27,12 @@ public class EmailTest extends AbstractOutputTest {
                         .finish()
         );
 
+        assertEquals(LocalDate.parse("2014-02-03"), email.getDomain().getFirstSeen());
         assertFalse(email.isDisposable());
         assertFalse(email.isFree());
         assertTrue(email.isHighRisk());
-        assertEquals(email.getFirstSeen(), "2017-01-02");
+        assertEquals("2017-01-02", email.getFirstSeen());
+        assertEquals(LocalDate.parse("2017-01-02"), email.getFirstSeenDate());
     }
 
     @Test
@@ -45,5 +52,6 @@ public class EmailTest extends AbstractOutputTest {
         assertFalse(email.isFree());
         assertTrue(email.isHighRisk());
         assertNull(email.getFirstSeen());
+        assertNull(email.getFirstSeenDate());
     }
 }