+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 CHANGELOG
 =========
 
+1.12.0
+-------------------
+
+* Added support for the new email output `/email/domain/first_seen` This can
+  be accessed via `response.getEmail().getDomain().getFirstSeen()`.
+* Added `Device.getLastSeenDateTime()`, `Device.getLocalDateTime()`,
+  `Email.getFirstSeenDate()`, and `GeoIp2Location.getLocalDateTime()` methods
+  that return `java.time` objects rather than strings.
+* The request event time is now stored internally as a `ZonedDateTime`. An
+  `Event.Builder.time(ZonedDateTime)` method was also added to the event
+  builder.
+
 1.11.0 (2020-02-21)
 -------------------